Create auto-updater for windows of service
When starts, the service will check if it's on the right version (can
be used to downgrade too). If it isn't, will download the new msi,
uninstall the old version based on the product code found at the
downloaded msi, which stops the service, and install the new version,
which restarts the service.