Change place where arguments are gotten
The args passed to OnStart are the ones passed on Windows Services
window. These ones are passed just once, in the next initialization,
and not saved. To save parameters for the service, one need to edit
windows registry, at
`HKEY_LOCAL_MACHINE\SYSTEM\CurrentControlSet\Services\BifrostService`
, key ImagePath, passing the parameters there. In this cas, the
parameters are passed to Program.Main, not to BifrostService.OnStart.